+++ daemons/raven-core/raven_core/tools/__init__.py
@@ -29,6 +29,7 @@ from . import time_tool
 from . import memory_tool
 from . import notify_tool
 from . import news_tool
+from . import finance_tool
 
 # Disabled until mesh integration:
 # from . import cerebras_tool   # second-LLM HTML generator, needs Flask sidecar
@@ -40,6 +41,7 @@ _TOOL_MODULES = [
     memory_tool,
     notify_tool,
     news_tool,
+    finance_tool,
 ]